+++ contracts/pns/IController.sol
@@ -16,17 +16,17 @@ interface IController {
     event NameRenewed(uint256 node, uint256 cost, uint256 expires, string name);
     event NameRegistered(address to, uint256 node, uint256 cost, uint256 expires, string name);
 
-    function nameRegisterByManager(string calldata name, address owner, uint256 duration, uint256 data, uint256[] calldata keyHashes, string[] calldata values) external returns(uint256);
+    function nameRegisterByManager(string calldata name, address owner, uint64 duration, uint256 data, uint256[] calldata keyHashes, string[] calldata values) external returns(uint256);
 
-    function nameRegister(string calldata name, address owner, uint256 duration) external payable returns(uint256);
+    function nameRegister(string calldata name, address owner, uint64 duration) external payable returns(uint256);
 
-    function nameRegisterWithConfig(string calldata name, address owner, uint256 duration, uint256 data, uint256[] calldata keyHashes, string[] calldata values) external payable returns(uint256);
+    function nameRegisterWithConfig(string calldata name, address owner, uint64 duration, uint256 data, uint256[] calldata keyHashes, string[] calldata values) external payable returns(uint256);
 
-    function renew(string calldata name, uint256 duration) external payable;
+    function renew(string calldata name, uint64 duration) external payable;
 
-    function renewByManager(string calldata name, uint256 duration) external;
+    function renewByManager(string calldata name, uint64 duration) external;
 
-    function nameRedeem(string calldata name, address owner, uint256 duration, uint256 deadline, bytes memory code) external returns(uint256);
+    function nameRedeem(string calldata name, address owner, uint64 duration, uint256 deadline, bytes memory code) external returns(uint256);
 
     function getTokenPrice() external view returns (int);
 
@@ -34,11 +34,11 @@ interface IController {
 
     function getPrices() view external returns(uint256[] memory, uint256[] memory);
 
-    function totalRegisterPrice(string memory name, uint256 duration) view external returns(uint256);
+    function totalRegisterPrice(string memory name, uint64 duration) view external returns(uint256);
 
-    function renewPrice(string memory name, uint256 duration) view external returns(uint256);
+    function renewPrice(string memory name, uint64 duration) view external returns(uint256);
 
     function basePrice(string memory name) view external returns(uint256);
 
-    function rentPrice(string memory name, uint256 duration) view external returns(uint256);
+    function rentPrice(string memory name, uint64 duration) view external returns(uint256);
 }
