+++ contracts/pns/PNS.sol
@@ -393,7 +393,7 @@ contract PNS is IPNS, IResolver, ERC721Upgradeable, ManagerOwnableUpgradeable, E
         emit MetadataUpdated(tokenIds);
     }
 
-    function register(string calldata name, address to, uint256 duration, uint256 baseNode) public override onlyManager returns(uint256) {
+    function register(string calldata name, address to, uint64 duration, uint256 baseNode) public override onlyManager returns(uint256) {
         // require(name.domainPreifxValid()); // skip due this will check by mintSubdomain
         uint256 tokenId = mintSubdomain(to, baseNode, name);
         require(available(tokenId), "tokenId not available");
@@ -406,11 +406,11 @@ contract PNS is IPNS, IResolver, ERC721Upgradeable, ManagerOwnableUpgradeable, E
         return tokenId;
     }
 
-    function renew(uint256 id, uint256 duration) public override onlyManager returns(uint256) {
+    function renew(uint256 id, uint64 duration) public override onlyManager returns(uint256) {
         require(records[id].origin == id, "not renewable");
         require(records[id].expire + duration + GRACE_PERIOD > block.timestamp + GRACE_PERIOD, "prevent overflow");
         // todo
-        records[id].expire += uint64(duration);
+        records[id].expire += duration;
         return records[id].expire;
     }
 }
